refactor(api): restructure FastAPI application and update entry point
