+++ api/Dockerfile
@@ -27,7 +27,7 @@ ENV PYTHONUNBUFFERED=1
 WORKDIR /app
 COPY . .
 EXPOSE 8000
-CMD ["python3", "-m", "uvicorn", "src.app:app", "--host=0.0.0.0", "--port=8000"]
+CMD ["python3", "-m", "uvicorn", "src.main:app", "--host=0.0.0.0", "--port=8000"]
 
 FROM development AS test
 ENV APP_ENV=test
